Fix ResultSet tests
